feat: power-up Step 5 — PluginHost interface (Gap F)
Narrow the plugin-registration surface to the minimum plugins need.
Previously `MontyPlugin.onRegister` received a full `MontyBridge`;
plugins had read/write access to the whole host API.

- Rename internal `PluginHost` class → `HostDispatch` (captures its
  actual role: dispatching host function calls).
- Introduce a public `PluginHost` interface exposing only
  `schemas` and `registerOs(OsCallHandler)` — the two things a
  plugin needs during registration.
- `MontyBridge implements PluginHost` so any bridge is-a host
  (no casts at call sites).
- `EventLoopPlugin` casts `host is DefaultMontyBridge` internally
  to reach `addStreamWrapper` — an implementation-detail escape
  hatch the plan explicitly permits.

All 437 tests pass, 13 skipped (integration-only).
