+++ lib/dart_monty_bridge.dart
@@ -27,7 +27,8 @@ export 'src/os_call/platform_handlers.dart';
 export 'src/os_call/sandboxed_fs_handler_stub.dart'
     if (dart.library.io) 'src/os_call/sandboxed_fs_handler.dart';
 export 'src/param_render_hint.dart';
-export 'src/plugin_host.dart' show MontyInterceptor;
+export 'src/host_dispatch.dart' show MontyInterceptor;
+export 'src/plugin_host.dart' show PluginHost;
 export 'src/plugin_registry.dart';
 export 'src/plugins/default_plugins.dart';
 export 'src/plugins/event_loop_plugin.dart';
