+++ lib/src/monty_runtime.dart
@@ -3,13 +3,13 @@ import 'dart:async';
 import 'package:dart_monty/src/bridge_event.dart';
 import 'package:dart_monty/src/bridge_logger.dart';
 import 'package:dart_monty/src/default_monty_bridge.dart';
+import 'package:dart_monty/src/host_dispatch.dart';
 import 'package:dart_monty/src/host_function.dart';
 import 'package:dart_monty/src/host_function_schema.dart';
 import 'package:dart_monty/src/monty_plugin.dart';
 import 'package:dart_monty/src/monty_runtime_ref.dart';
 import 'package:dart_monty/src/monty_runtime_state.dart';
 import 'package:dart_monty/src/os_call/os_handlers.dart';
-import 'package:dart_monty/src/plugin_host.dart';
 import 'package:dart_monty/src/plugin_registry.dart';
 import 'package:dart_monty/src/tool_surface.dart';
 import 'package:dart_monty_core/dart_monty_core.dart';
